+++ src/services/item/plugins/etherpad/test/service-api.test.ts
@@ -7,7 +7,7 @@ import waitForExpect from 'wait-for-expect';
 
 import { FastifyInstance } from 'fastify';
 
-import { EtherpadItemType, HttpMethod, PermissionLevel } from '@graasp/sdk';
+import { EtherpadItemType, HttpMethod, ItemType, PermissionLevel } from '@graasp/sdk';
 
 import build, { clearDatabase } from '../../../../../../test/app';
 import { ETHERPAD_PUBLIC_URL } from '../../../../../utils/config';
@@ -18,6 +18,7 @@ import {
 } from '../../../../itemMembership/test/fixtures/memberships';
 import { Member } from '../../../../member/entities/member';
 import { BOB, saveMember } from '../../../../member/test/fixtures/members';
+import { Item } from '../../../entities/Item';
 import { ItemRepository } from '../../../repository';
 import { MAX_SESSIONS_IN_COOKIE } from '../constants';
 import { EtherpadItemService } from '../service';
@@ -259,6 +260,7 @@ describe('Etherpad service API', () => {
         member: bob,
         item: {
           name: "bob's test etherpad item",
+          type: ItemType.ETHERPAD,
           extra: EtherpadItemService.buildEtherpadExtra({
             groupID: MOCK_GROUP_ID,
             padName: MOCK_PAD_NAME,
@@ -477,7 +479,7 @@ describe('Etherpad service API', () => {
       };
       expect(name).toEqual('sessionID');
       const sessions = value.split(',');
-      expect(sessions.length).toEqual(MAX_SESSIONS_IN_COOKIE);
+      expect([MAX_SESSIONS_IN_COOKIE, MAX_SESSIONS_IN_COOKIE - 1]).toContain(sessions.length);
       // the first (oldest) session should not be in the cookie
       Array.from(
         { length: MAX_SESSIONS_IN_COOKIE - 1 },
@@ -732,7 +734,7 @@ describe('Etherpad service API', () => {
 
       const copy = (await ItemRepository.findOneBy({
         id: And(Not(item.id), Not(parent.item.id)),
-      })) as EtherpadItemType;
+      })) as Item<typeof ItemType.ETHERPAD>;
       expect(copy).not.toBeNull();
 
       const { createGroupIfNotExistsFor, copyPad } = await reqsParams;