+++ src/services/item/plugins/published/entities/itemPublished.ts
@@ -11,15 +11,13 @@ import {
 } from 'typeorm';
 import { v4 } from 'uuid';
 
-import { ItemPublished as GraaspItemPublished } from '@graasp/sdk';
-
 import { Member } from '../../../../member/entities/member';
 import { Item } from '../../../entities/Item';
 
 @Entity()
 @Index('IDX_gist_item_published_path', { synchronize: false })
 @Unique('published-item', ['item'])
-export class ItemPublished extends BaseEntity implements GraaspItemPublished {
+export class ItemPublished extends BaseEntity {
   @PrimaryGeneratedColumn('uuid')
   id: string = v4();