+++ src/utils/errors.ts
@@ -327,8 +327,7 @@ export class EmptyCurrentPassword extends CoreError {
 }
 
 export class UnauthorizedMember extends CoreError {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  constructor(data?: unknown) {
+  constructor(_data?: unknown) {
     // this status code is custom for the browser to know it needs to refresh its token
     super({
       code: 'GERR027',